% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_iNEXT.R
\name{do_iNEXT}
\alias{do_iNEXT}
\title{Generate iNEXT Sample-Size-Based Richness Plot
Author: Jacob Agerbo Rasmussen}
\usage{
do_iNEXT(
  data,
  sample_data = NULL,
  group_var = NULL,
  table = FALSE,
  threads = 2
)
}
\arguments{
\item{data}{The dataset containing the observations}

\item{sample_data}{The dataset containing the sample data (default: NULL)}

\item{group_var}{The variable used to group the samples (default: NULL)}

\item{table}{A logical value indicating whether to return the extrapolated data as a table (default: FALSE)}

\item{threads}{The number of parallel threads to use (default: 2)}
}
\description{
The \code{do_iNEXT} function generates a sample-size-based richness plot using the iNEXT package.
The function takes the following parameters:
}
\details{
\itemize{
\item \code{data}: The dataset containing the observations.
\item \code{sample_data}: The dataset containing the sample data (default: NULL).
\item \code{group_var}: The variable used to group the samples (default: NULL).
\item \code{table}: A logical value indicating whether to return the extrapolated data as a table (default: FALSE).
\item \code{threads}: The number of parallel threads to use (default: 2).
}

The function loads the required packages (\code{tidyverse}, \code{iNEXT}, \code{ggtext}, \code{ggpubr}, \code{doParallel}, and \code{future}) and calculates iNEXT diversity for the provided data.
If a group variable is provided, the function aggregates the data by summing values across each row and performs iNEXT analysis on the relevant columns of the aggregated data.
If no group variable is provided, the function sums the values across the columns and performs iNEXT analysis on the resulting data.

The function then generates an iNEXT plot type 1 with reordered facets, sets the theme to minimal, and adds a dashed line representing the intercept and slope.
The plot title, y-axis label, and caption are customized using the \code{ggtext} package.

If the \code{table} parameter is set to \code{TRUE}, the function returns both the plot and an extrapolated data table as a list. Otherwise, it returns only the plot.

Finally, the function stops the parallel cluster.

Example usage:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Load data
data(iris)

# Generate iNEXT plot for iris dataset
do_iNEXT(iris[,1:4])
}\if{html}{\out{</div>}}
}
