% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Power_analysis.R
\name{do_Power}
\alias{do_Power}
\title{Perform Power Analysis for Sample Size and Effect Size
Author: Jacob Agerbo Rasmussen
The \code{do_Power} function is used to perform a power analysis for sample size and effect size. It calculates the required sample size for a given effect size, or the required effect size for a given sample size, based on the desired significance level and power.}
\usage{
do_Power(
  max_sample_size = 100,
  min_effect_size = 0.01,
  sig.level = 0.05,
  power = 0.8
)
}
\arguments{
\item{max_sample_size}{The maximum sample size to consider in the power analysis.}

\item{min_effect_size}{The minimum effect size to consider in the power analysis.}

\item{sig.level}{The desired significance level (Type I error rate) for the power analysis.}

\item{power}{The desired power (1 - Type II error rate) for the power analysis.}
}
\value{
A combined scatter plot of sample size versus effect size and effect size versus sample size.
}
\description{
The function performs the following steps:
\enumerate{
\item Loads the necessary packages, including \code{pwr}, \code{tidyverse}, \code{scales}, and \code{cowplot}.
\item Checks if the maximum sample size is less than 20, and displays a warning message if it is.
\item Defines the effect size categories (low, medium, and large).
\item Calculates the required sample sizes for different effect sizes using the \code{pwr.t.test} function.
\item Creates a scatter plot of sample size versus effect size, with a logarithmic x-axis.
\item Adds horizontal lines and text annotations to indicate the effect size categories.
\item Calculates the required effect sizes for different sample sizes using the \code{pwr.t.test} function.
\item Creates a second scatter plot of effect size versus sample size, with a logarithmic x-axis.
\item Adds horizontal lines and text annotations to indicate the effect size categories.
\item Combines both scatter plots into a single plot using \code{cowplot::plot_grid}.
\item Returns the combined plot as the output.
}
}
\details{
Example usage:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Perform power analysis
do_Power(max_sample_size = 100, min_effect_size = 0.01, sig.level = 0.05, power = 0.8)
}\if{html}{\out{</div>}}
}
