% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_boxplot.R
\name{generate_boxplot}
\alias{generate_boxplot}
\title{Generate a boxplot
Author: Jacob Agerbo Rasmussen}
\usage{
generate_boxplot(
  data,
  sample_data = NULL,
  group_var = NULL,
  plot_title = NULL,
  palette = "Dark2",
  violin = FALSE,
  jitter = FALSE,
  test = FALSE
)
}
\arguments{
\item{data}{The dataset to be analyzed}

\item{sample_data}{The sample information dataset}

\item{group_var}{The grouping variable to be used for generating a grouped boxplot}

\item{plot_title}{The title of the plot}

\item{palette}{The color palette to be used for the plot}

\item{violin}{A logical value indicating whether to generate a violin plot instead of a boxplot}

\item{jitter}{A logical value indicating whether to add jitter points to the plot}

\item{test}{A logical value indicating whether to perform a statistical test and add p-values and significance labels to the plot}
}
\value{
A boxplot or violin plot
}
\description{
The \code{generate_boxplot} function is used for generating a boxplot of a provided dataset.
If \code{sample_data} and \code{group_var} are specified, the function generates a grouped boxplot showing the distribution of values in each group.
If \code{violin} is set to TRUE, the function generates a violin plot instead of a boxplot.
If \code{jitter} is set to TRUE, the function adds jitter points to the plot.
If \code{test} is set to TRUE, the function performs a statistical test (Levene's test) and adds p-values and significance labels to the plot.
}
\examples{
# Generate a simple boxplot
generate_boxplot(data, group_var = group_var, jitter = TRUE,
               violin = FALSE, test = TRUE)
}
