% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overview.R
\name{get_overview}
\alias{get_overview}
\title{Generate descriptive plots of a provided dataset
Author: Jacob Agerbo Rasmussen}
\usage{
get_overview(data = data, sample_data = sample_data, group_var = NULL)
}
\arguments{
\item{data}{The dataset to be analyzed}

\item{sample_data}{The sample information dataset}

\item{group_var}{The variable used for grouping the data}
}
\value{
A list of plots
}
\description{
The \code{get_overview} function is used for generating descriptive plots of a provided dataset.
If \code{group_var} is specified, the function generates four plots: a ridgeline plot, an empirical cumulative distribution function (ECDF) plot, a Q-Q plot, and a distribution plot of individual features.
If \code{group_var} is not specified, the function generates a distribution plot of individual features.
}
