% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_barplot.R
\name{make_barplot}
\alias{make_barplot}
\title{Generate a barplot of a provided dataset
Author: Jacob Agerbo Rasmussen}
\usage{
make_barplot(
  data = data,
  sample_data = NULL,
  tax_data = NULL,
  taxa = NULL,
  group = NULL,
  grouping = FALSE
)
}
\arguments{
\item{data}{The dataset to be analyzed}

\item{sample_data}{The sample information dataset}

\item{tax_data}{The taxonomic information dataset}

\item{taxa}{The taxonomic group to be plotted}

\item{group}{The grouping variable to be used for generating a grouped barplot}

\item{grouping}{A logical value indicating whether to generate a grouped barplot}
}
\value{
A ggplot barplot
}
\description{
The \code{make_barplot} function is used for generating a barplot of a provided dataset.
If \code{tax_data} and \code{taxa} are specified, the function generates a stacked barplot showing the relative abundance of each taxonomic group in each sample.
If \code{grouping} is set to TRUE, the function generates a grouped barplot showing the mean relative abundance of each taxonomic group in each sample group.
If \code{tax_data} is not specified, the function generates a simple barplot showing the values of each variable in each sample.
}
