% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lme.R
\name{make_lme_cor}
\alias{make_lme_cor}
\title{Perform Linear Mixed-Effects Modeling and Correlation Analysis
Author: Jacob Agerbo Rasmussen}
\usage{
make_lme_cor(
  data,
  sample_data,
  feature,
  exp_var,
  random_var,
  datatype = c("logcpm", "relabu", "counts"),
  tolerance
)
}
\arguments{
\item{data}{A matrix or data frame containing the data for the analysis}

\item{sample_data}{A data frame containing sample information such as treatment and group}

\item{feature}{The name of the feature (dependent variable) to be analyzed}

\item{exp_var}{The name of the explanatory variable}

\item{random_var}{The name of the random effect variable}

\item{datatype}{The type of data to be analyzed (default: "logcpm")}

\item{tolerance}{The tolerance value used for computing adjusted correlation}
}
\value{
A list containing the scatter plot and statistics
}
\description{
The \code{make_lme_cor} function is used to perform linear mixed-effects modeling and correlation analysis.
It takes the following parameters:
}
\details{
\itemize{
\item \code{data}: A matrix or data frame containing the data for the analysis.
\item \code{sample_data}: A data frame containing sample information such as treatment and group.
\item \code{feature}: The name of the feature (dependent variable) to be analyzed.
\item \code{exp_var}: The name of the explanatory variable.
\item \code{random_var}: The name of the random effect variable.
\item \code{datatype}: The type of data to be analyzed. It can be "logcpm", "relabu", or "counts".
\item \code{tolerance}: The tolerance value used for computing adjusted correlation.
}

The function performs the following steps:
\enumerate{
\item Loads required packages (\code{tidyverse}, \code{performance}, \code{ggpubr}, \code{animalcules}, \code{lmerTest}, \code{broom.mixed}, and \code{ggplot2}).
\item Prepares the data by transforming it based on the specified data type.
\item Removes rows with zero counts, if any.
\item Selects the relevant columns from the data frame.
\item Creates a scatter plot with a linear regression line for each group.
\item Estimates a mixed-effects model using the \code{lmerTest} package.
\item Retrieves the sign of the coefficient and the adjusted correlation.
\item Computes the p-value using \code{anova} with Kenward-Roger degrees of freedom adjustment.
\item Returns the plot and statistics as a list.
}

Example usage:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Load data
data <- read.csv("data.csv")
sample_data <- read.csv("sample_info.csv")

# Perform linear mixed-effects modeling and correlation analysis
make_lme_cor(data, sample_data, "feature", "exp_var", "random_var", datatype = "logcpm", tolerance = 0.1)
}\if{html}{\out{</div>}}
}
