% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_test_data.R
\name{make_test_data}
\alias{make_test_data}
\title{Make Test Data
Author: Jacob Agerbo Rasmussen}
\usage{
make_test_data(no_samples = 100)
}
\arguments{
\item{no_samples}{The number of samples to generate. The default is 100.}
}
\value{
A list containing the generated data and sample information.
}
\description{
The \code{make_test_data} function generates test data for use in various analyses. It creates a matrix of random numbers using the Poisson distribution and combines them into several data sets of different sizes and variances. It also generates sample information such as group variable, batch variable, treatment variable, and more, and stores them in a data frame.
}
\details{
The function performs the following steps:
\enumerate{
\item Sets the seed for reproducibility.
\item Generates a matrix of random numbers using the Poisson distribution.
\item Combines the matrix into several data sets of different sizes and variances.
\item Generates sample information such as group variable, batch variable, treatment variable, and more, and stores them in a data frame.
\item Prints a message indicating that test data has been generated.
\item Returns a list containing the generated data and sample information.
}

Example usage:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Generate test data
test_data <- make_test_data()

# Access generated data and sample information
data <- test_data$data
sample_data <- test_data$sample_data
}\if{html}{\out{</div>}}
}
