% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess Data}
\usage{
preprocess_data(
  data,
  sample_data = NULL,
  batch = "Group",
  clean_data = TRUE,
  sum_scaling = TRUE,
  magic_norm = FALSE,
  correct_batch = TRUE
)
}
\arguments{
\item{data}{A matrix or data frame containing the input data to be preprocessed.}

\item{sample_data}{A data frame containing sample information, such as batch information.}

\item{batch}{The name of the column in \code{sample_data} that represents the batch information.}

\item{clean_data}{A logical value indicating whether to perform data cleaning (removing missing or invalid values and duplicate samples or features). Default is \code{TRUE}.}

\item{sum_scaling}{A logical value indicating whether to perform total sum scaling normalization. Default is \code{TRUE}.}

\item{magic_norm}{A logical value indicating whether to perform inverse normal transformation (INT) for normalization of continuous phenotype data. Default is \code{FALSE}.}

\item{correct_batch}{A logical value indicating whether to perform batch effect correction using ComBat. Default is \code{TRUE}.}
}
\value{
A list containing the combined plot, processed data, and raw data.
}
\description{
The \code{preprocess_data} function is used to preprocess the input data before further analysis. It performs several preprocessing steps, including data cleaning, normalization, and batch effect correction using ComBat.
}
\details{
The function performs the following steps:
\enumerate{
\item Loads the necessary packages, including \code{sva}, \code{tidyverse}, and \code{cowplot}.
\item Stores a copy of the original data in \code{prior_data}.
\item If \code{clean_data} is set to \code{TRUE}, removes missing or invalid values from the data and removes duplicate samples or features.
\item If \code{sum_scaling} is set to \code{TRUE}, performs total sum scaling normalization on the data.
\item If \code{magic_norm} is set to \code{TRUE}, applies inverse normal transformation (INT) to normalize continuous phenotype data.
\item If \code{correct_batch} is set to \code{TRUE}, identifies batch information from \code{sample_data} and performs batch effect correction using ComBat.
\item Generates outlier detection plots for raw and processed data if applicable.
\item Generates PCA plots for raw and batch-corrected data if applicable.
\item Combines the outlier detection plots and PCA plots into a single plot.
\item Returns the combined plot, processed data, and raw data as output.
}

Example usage:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Preprocess data
preprocessed <- preprocess_data(data, sample_data = sample_data, batch = "Group", clean_data = TRUE, sum_scaling = TRUE, magic_norm = FALSE, correct_batch = TRUE)

# Access the combined plot
preprocessed$plot

# Access the processed data
preprocessed$processed_data

# Access the raw data
preprocessed$raw_data
}\if{html}{\out{</div>}}
}
